+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ### Added
 
+- Add `final_reference` field in `Invoice` model
 - Update Volta configuration with yarn fixed version
 - Allow custom discount (percentage or fixed amount) when creating admin orders
 - Introduce django-waffle and expose its status through `waffle_status`